Use Ansible to create review permissions.
In order for reviewing to work correctly, permissions need to be added
for each conference section. Create an Ansible task to run the
management command that creates these permissions.

It should be okay to run this command on every deployment. If the slug
of a conference section changes, it will be necessary to run the command
(either manually or part of a deployment) and then update the
permissions of the reviewing team(s).