Pin symposion checkout to specific commit.
Pin our version of symposion to a specific commit instead of the patched
branch. This makes it easier to update conference sites with newer
versions (since the virtualenv creation portion of deployment relies on
Git commit hashes to determine whether a new virtualenv needs to be
created, a update to the patched branch will not cause symposion on
conference sites to be updated if there is no corresponding change in
the conf_site repository).