Change speaker profile link in footer.
Since the speaker profile page is only available if the speaker has
accepted presentations, it does not make sense to link to it in the
footer (since all non-staff users will receive a 404 until proposals are
reviewed).

Change link to speaker profile edit page. Move link below links to
dashboard and admin, since it is arguably less important.