Change where STATICFILES_STORAGE is defined.
Move STATICFILES_STORAGE from production settings to base settings.
While ManifestStaticFilesStorage does not run when DEBUG = True, this
still makes it easier to override the static storage backend in local
development, and we should try to have as few settings defined in the
environment-specific settings files as possible.

Note that ManifestStaticFilesStorage does not work in tests, so we need
to explicitly disable it in Travis-CI's settings.