Update presentations when proposals are saved.
Overload Proposals' save method so that any associated Presentation will
be updated as well. This should make it easier for conference speakers
to update the information on their presentation page after their
proposal has been accepted.

It is possible that this could be moved to ProposalBase and sent
upstream to Symposion.