Fix ability to click on venue page Google map.
Ensure that users can interact with the embedded Google map by ensuring
that the hotel information section does not appear on top of it. This
was happening because the hotel information section used the "sec1" ID,
which has a higher z-index than default.

Fix this by removing the troublesome ID from both the venue and hotel
information sections.