Fix pagination styling on reviews screen.
Symposion uses an external JQuery plugin called DataTables
(https://datatables.net/) to create and manage the tables used in the
reviews section. At some unknown point in the past, the pagination
navigation stopped being styled correctly (Symposion's DataTables
applies the "pagination" class to a div element, when it should be
applied to the ul that contains links to the individual pages).

Since attempting to upgrade DataTables might cause other issues, we just
add the relevant class to the ul directly after the table has been
created.