Update symposion to change Proposal field labels.
Update symposion to change field labels for the Proposal model's
description and abstract fields.

It is necessary to make this change in symposion because the two fields
are defined in symposion's ProposalBase model. They can't be overwritten
in the conference site codebase because ProposalBase isn't an abstract
model and we're not using Django 1.10+.