Change dialect used to create exported CSV files.
Change the dialect used when creating exported CSV files from the
default (`excel`) to `unix_dialect`. Since the latter quotes all fields
and uses Unix line endings, it is a better dialect.

This should also (hopefully) fix downloading issues on production. CSV
files with the `excel` dialect have Windows line endings. When an
exported CSV file is downloaded, the server changes these line endings
to Unix. This changes the size of the downloaded file and invalidates
the Content-Length HTTP header, which causes the download to fail over
HTTP/2.