Coerce ProposalVote.score into an integer.
Ensure that form-submitted vote score is an integer by coercing it in
the POSTing view. Otherwise, we will cache the score's value (e.g. "3")
instead of its textual representation (e.g. "+1").

This commit also changes ProposalVoteForm to use a TypedChoiceField
(which will prevent a regression of this issue if we change how saving
votes works in the future).